// src/app/shared/models/appUserDtos/updateAppUserDto.ts
export interface UpdateAppUserDto {
    id: string;
    firstName: string | null;
    lastName: string | null;
    introduction: string | null;
    interest: string | null;
    idealType: string | null;
    address: string | null;
    ward: string | null;
    district: string | null;
    city: string | null;
    role: string | null;
}